requirements.txt
Flask==2.3.3
google-cloud-storage==2.14.0
google-cloud-vision==3.4.0
google-generativeai
